% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalModel.R
\docType{class}
\name{FunctionalModel-class}
\alias{FunctionalModel-class}
\alias{FunctionalModel}
\title{A FunctionalModel for a Functional Model}
\description{
This class holds the a blueprint for a functional model, i.e.,
  an unparameterized function model. Such a model is defined by a
  function \code{f} which accepts one scalar input \code{x} and a
  parameterization vector \code{par} and returns an output scalar \code{y}.
  The model depends on the parameterization \code{par}, which will later be
  subject to optimization to make the function \code{f(x, par)} fit to a
  model dataset \code{(x, y)}.
}
\section{Slots}{

\describe{
\item{\code{f}}{the model function, taking as parameters a value \code{x} followed by
a parameter vector \code{par}}

\item{\code{estimator}}{is a function which takes in a vector of \code{x} and a
vector of \code{y} values as well as limits \code{paramLower} and
\code{paramUpper} and returns an estimate of the parameters, or \code{NULL}
if no estimate can be made better than just standard random numbers}

\item{\code{gradient}}{a function which takes in a value \code{x} and \code{par} and
returns a vector with the gradient for each parameter dimension}

\item{\code{paramCount}}{the number of model parameters}

\item{\code{paramLower}}{the lower bounds for the parameters, or \code{NULL} if none
are required. An element of the vector may be set of \code{NA} if no lower
limit for that limit is specified (while lower limits are given for other
parameter values).}

\item{\code{paramUpper}}{the upper bounds for the parameters, or \code{NULL} if none
are required. An element of the vector may be set of \code{NA} if no lower
upper for that limit is specified (while upper limits are given for other
parameter values).}
}}

\seealso{
FunctionalModel.new
}
