% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{FunctionalModel.all}
\alias{FunctionalModel.all}
\title{Obtain a List with the All Functional Models Provided by this Library}
\usage{
FunctionalModel.all()
}
\value{
a list with all models provided by this library.
}
\description{
This function returns a list with the all the models provided by
  this library. Currently, these comprise standard constant, linear,
  quadratic, and cubic models as well as several sigmoid-style models such as
  exponential decay, logistic models, the Gompertz model, and an exponential
  log-linear model.
}
\seealso{
FunctionalModel.new

FunctionalModel.linear

FunctionalModel.quadratic
}
