% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{FunctionalModel.constant}
\alias{FunctionalModel.constant}
\title{Obtain the Simple Constant Model \code{y = f(x) = a}}
\usage{
FunctionalModel.constant()
}
\description{
A simple constant model, i.e., a model of the form \code{y =
  f(x) = a} with one parameter (\code{a}).
}
