% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic.R
\name{FunctionalModel.cubic.from.four.points}
\alias{FunctionalModel.cubic.from.four.points}
\title{Compute the parameter values of a Cubic Function from four Point
  Coordinates}
\usage{
FunctionalModel.cubic.from.four.points(x1, y1, x2, y2, x3, y3, x4, y4)
}
\arguments{
\item{x1}{the first \code{x}-coordinate}

\item{y1}{the first \code{y}-coordinate}

\item{x2}{the second \code{x}-coordinate}

\item{y2}{the second \code{y}-coordinate}

\item{x3}{the third \code{x}-coordinate}

\item{y3}{the third \code{y}-coordinate}

\item{x4}{the fourth \code{x}-coordinate}

\item{y4}{the fourth \code{y}-coordinate}
}
\value{
a vector of type \code{(a,b,c,d)}, such than
  \code{f(x)=a+b*x+c*x^2+d*x^3} or \code{NULL} if no finite result is
  possible
}
\description{
This function returns the four parameters of a cubic model from
  four point coordinates.
}
