% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic.R
\name{FunctionalModel.cubic}
\alias{FunctionalModel.cubic}
\title{Obtain the Simple Cubic Model \code{y = f(x) = a + b*x + c*x^2 + d*x^3}}
\usage{
FunctionalModel.cubic()
}
\description{
A simple cubic model, i.e., a model of the form \code{y =
  f(x) = a + b*x + c*x^2 + d*x^3} with four parameters (\code{a}, \code{b}, \code{c}, and
  \code{4})).
}
