% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expLogLinear.R
\name{FunctionalModel.expLogLinear.1}
\alias{FunctionalModel.expLogLinear.1}
\title{Obtain the First Variant of the Exponential Log-Linear Model}
\usage{
FunctionalModel.expLogLinear.1()
}
\description{
This function returns the first variant of the generalized
  exponential log-linear model \code{a + b*exp(c*log(x+d))} where \code{b} is
  enforced to be positive and \code{c} is enforced to be negative.
}
\seealso{
FunctionalModel.expLogLinear.2
}
