% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expLogLinear.R
\name{FunctionalModel.expLogLinear.2}
\alias{FunctionalModel.expLogLinear.2}
\title{Obtain the Second Variant of the Exponential Log-Linear Model}
\usage{
FunctionalModel.expLogLinear.2()
}
\description{
This function returns the first variant of the generalized
  exponential log-linear model \code{a + b*exp(c*log(x+d))} where \code{b} is
  enforced to be negative and \code{c} is enforced to be positive.
}
\seealso{
FunctionalModel.expLogLinear.1
}
