% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentialDecay.R
\name{FunctionalModel.exponentialDecay.2}
\alias{FunctionalModel.exponentialDecay.2}
\title{Obtain the Second Variant of the Generalized Exponential Decay Model}
\usage{
FunctionalModel.exponentialDecay.2()
}
\description{
This function returns the first variant of the generalized
  exponential decay model \code{a+(b*exp(c*x^d)} where both \code{b} and
  \code{d} are enforced to be negative.
}
\seealso{
FunctionalModel.exponentialDecay.1
}
