% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{FunctionalModel.gompertz.1}
\alias{FunctionalModel.gompertz.1}
\title{Obtain the First Variant of the Gompertz Model}
\usage{
FunctionalModel.gompertz.1()
}
\description{
This function returns the first variant of the generalized
  Gompertz model \code{a+(b*exp(c*exp(d*x)))} where both \code{b} and
  \code{d} are enforced to be positive.
}
\seealso{
FunctionalModel.gompertz.2
}
