% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{FunctionalModel.gompertz.2}
\alias{FunctionalModel.gompertz.2}
\title{Obtain the Second Variant of the Gompertz Model}
\usage{
FunctionalModel.gompertz.2()
}
\description{
This function returns the first variant of the generalized
  Gompertz model \code{a+(b*exp(c*exp(d*x)))} where both \code{b} and
  \code{d} are enforced to be negative.
}
\seealso{
FunctionalModel.gompertz.1
}
