% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{FunctionalModel.linear.from.two.points}
\alias{FunctionalModel.linear.from.two.points}
\title{Compute the parameter values of a Linear Function from two Point
  Coordinates}
\usage{
FunctionalModel.linear.from.two.points(x1, y1, x2, y2)
}
\arguments{
\item{x1}{the first \code{x}-coordinate}

\item{y1}{the first \code{y}-coordinate}

\item{x2}{the second \code{x}-coordinate}

\item{y2}{the second \code{y}-coordinate}
}
\value{
a vector of type \code{(m,n)}, such than \code{f(x)=n+m*x} or
  \code{NULL} if no finite result is possible
}
\description{
This function returns the two parameters of a linear model from
  two point coordinates.
}
