% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{FunctionalModel.linear}
\alias{FunctionalModel.linear}
\title{Obtain the Simple Linear Model \code{y = f(x) = a+b*x}}
\usage{
FunctionalModel.linear()
}
\description{
A simple linear model, i.e., a model of the form \code{y = f(x)
  = a+b*x} with two parameters (\code{a} and \code{b})).
}
