% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{FunctionalModel.logistic.1}
\alias{FunctionalModel.logistic.1}
\title{Obtain the First Variant of the Logistic Model}
\usage{
FunctionalModel.logistic.1()
}
\description{
This function returns the first variant of the logistic model
  \code{a+b/(1+c*x^d)} where both \code{b} and \code{d} are enforced to be
  positive.
}
\seealso{
FunctionalModel.logistic.2
}
