% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{FunctionalModel.logistic.2}
\alias{FunctionalModel.logistic.2}
\title{Obtain the Second Variant of the Logistic Model}
\usage{
FunctionalModel.logistic.2()
}
\description{
This function returns the first variant of the logistic model
  \code{a+b/(1+c*x^d)} where both \code{b} and \code{d} are enforced to be
  negative.
}
\seealso{
FunctionalModel.logistic.1
}
