% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpDefault.R
\name{FunctionalModel.mlp.default}
\alias{FunctionalModel.mlp.default}
\title{Get the List of Default MLP-based Models}
\usage{
FunctionalModel.mlp.default()
}
\value{
the list of default MLPs
}
\description{
Obtain a list of MLP-based models which are reasonable for a
  wide variety of applications.
}
\seealso{
FunctionalModel.mlp.monotonous
}
