% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpDefault.R
\name{FunctionalModel.mlp.monotonous}
\alias{FunctionalModel.mlp.monotonous}
\title{Get the List of Default Monotonously Decreasing MLP-based Models}
\usage{
FunctionalModel.mlp.monotonous()
}
\value{
the list of monotonous MLPs
}
\description{
Obtain a list of monotonously decreasing MLP-based models which
  are reasonable for a wide variety of applications.
}
\seealso{
FunctionalModel.mlp.default
}
