% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{FunctionalModel.monotonous}
\alias{FunctionalModel.monotonous}
\title{Obtain a List with the Default Monotonously Decreasing Functional
  Models Provided by this Library}
\usage{
FunctionalModel.monotonous()
}
\value{
a list with the default models provided by this library.
}
\description{
This function returns a list with the default monotonously
  decreasing models provided by this library. Currently, these comprise
  standard constant and linear model as well as several sigmoid-style models
  such as exponential decay, logistic models, the Gompertz model, and an
  exponential log-linear model.
}
\seealso{
FunctionalModel.new

FunctionalModel.constant

FunctionalModel.linear

FunctionalModel.exponentialDecay.1

FunctionalModel.exponentialDecay.2

FunctionalModel.logistic.1

FunctionalModel.logistic.2

FunctionalModel.gompertz.1

FunctionalModel.gompertz.2

FunctionalModel.expLogLinear.1

FunctionalModel.expLogLinear.2
}
