% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalModel.R
\name{FunctionalModel.new}
\alias{FunctionalModel.new}
\title{Create a new instance of \code{\link{FunctionalModel}}}
\usage{
FunctionalModel.new(f, paramCount, gradient = NULL, estimator = NULL,
  paramLower = NULL, paramUpper = NULL, name = functionToString(f))
}
\arguments{
\item{f}{the model function, taking as parameters a value \code{x} followed
by a parameter vector \code{par}}

\item{paramCount}{the number of model parameters}

\item{gradient}{a function which takes in a value \code{x} and \code{par} and
returns a vector with the gradient for each parameter dimension}

\item{estimator}{is a function which takes in a vector of \code{x} and a
vector of \code{y} values as well as limits \code{paramLower} and
\code{paramUpper} and returns an estimate of the parameters, or \code{NULL}
if no estimate can be made better than just standard normal random numbers}

\item{paramLower}{the lower bounds for the parameters, or \code{NULL} if none
are required. An element of the vector may be set of \code{NA} if no lower
limit for that limit is specified (while lower limits are given for other
parameter values).}

\item{paramUpper}{the upper bounds for the parameters, or \code{NULL} if none
are required. An element of the vector may be set of \code{NA} if no lower
upper for that limit is specified (while upper limits are given for other
parameter values).}

\item{name}{the name of the model}
}
\value{
the new functional functional model
}
\description{
Instantiate the class \code{\link{FunctionalModel}}.
  Negative infinite lower limits will be aliased to \code{NA}. If all lower
  limits are \code{NA}, the lower limits vector will be set to \code{NULL}.
  Alias all positive infinite upper limits to \code{NA}. If all upper limits
  are \code{NA}, the vector of upper limits will be set to \code{NULL}.
}
