% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic.R
\name{FunctionalModel.quadratic.from.three.points}
\alias{FunctionalModel.quadratic.from.three.points}
\title{Compute the parameter values of a Quadratic Function from three Point
  Coordinates}
\usage{
FunctionalModel.quadratic.from.three.points(x1, y1, x2, y2, x3, y3)
}
\arguments{
\item{x1}{the first \code{x}-coordinate}

\item{y1}{the first \code{y}-coordinate}

\item{x2}{the second \code{x}-coordinate}

\item{y2}{the second \code{y}-coordinate}

\item{x3}{the third \code{x}-coordinate}

\item{y3}{the third \code{y}-coordinate}
}
\value{
a vector of type \code{(a,b,c)}, such than \code{f(x)=a+b*x+c*x^2} or
  \code{NULL} if no finite result is possible
}
\description{
This function returns the three parameters of a quadratic model
  from three point coordinates.
}
