% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parEstimate.R
\name{par.estimate}
\alias{par.estimate}
\title{Create a Parameter Estimate for a Given Functional Model}
\usage{
par.estimate(functionalModel, x = NULL, y = NULL, par = NULL)
}
\arguments{
\item{functionalModel}{the model function functional model}

\item{x}{the input values}

\item{y}{the corresponding output values}

\item{par}{an existing estimate, or \code{NULL} if no previous estimate exists}
}
\value{
the parameter estimate
}
\description{
Use the data available in the functional functional model
  together with the \code{x} and \code{y} coordinates to be modeled to
  produce an initial first guess of the parameters. This method should not be
  confused with actually fitting the model. It won't do that, it just
  provides a reasonable initial guess. It therefore uses the estimator
  function provided by the functional model, if any, and then attempts to fix
  any parameter values violating the constraints.
}
