% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic.R
\name{quadratic}
\alias{quadratic}
\title{Obtain the Simple Quadratic Model \code{y = f(x) = a + b*x + c*x^2}}
\usage{
quadratic()
}
\description{
A simple quadratic model, i.e., a model of the form \code{y = f(x)
  = a + b*x + c*x^2} with three parameters (\code{a}, \code{b}, and \code{c})).
}
