% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{FunctionalModel.mlp}
\alias{FunctionalModel.mlp}
\title{Create a Multi-Layer Perceptron Model with the Given Layers and
Activation Function}
\usage{
FunctionalModel.mlp(layers, func = "tanh")
}
\arguments{
\item{layers}{the vector of layers, e.g., \code{c(1)} for a perceptron with a
single input=output node, \code{1, 2, 1} with an input node, two hidden
nodes, and an output node.}

\item{func}{the activation function}
}
\value{
a functional model representing the given perceptron
}
\description{
With this function, we can create a multi-layer perceptron model
  with the specified \code{layers} configuration and activation
  \code{func}tion. Each neuron has a bias parameter plus \code{n} weights for
  all incoming connections from the previous layer. The output of the last
  layer is also weighted and added to a bias parameter.
}
